+++ how-to/workspace-platform-starter/client/src/framework/platform/platform-override.ts
@@ -31,12 +31,15 @@ import { applyClientSnapshot, decorateSnapshot } from "../snapshot-source";
 import { setCurrentColorSchemeMode } from "../themes";
 import { isEmpty } from "../utils";
 import { getAllVisibleWindows, getPageBounds } from "./browser";
+import { cleanupPage, cleanupWorkspace } from "./cleanup";
 import { closedown as closedownPlatform } from "./platform";
 
+const WORKSPACE_ENDPOINT_ID_LIST = "workspace-list";
 const WORKSPACE_ENDPOINT_ID_GET = "workspace-get";
 const WORKSPACE_ENDPOINT_ID_SET = "workspace-set";
 const WORKSPACE_ENDPOINT_ID_REMOVE = "workspace-remove";
 
+const PAGE_ENDPOINT_ID_LIST = "page-list";
 const PAGE_ENDPOINT_ID_GET = "page-get";
 const PAGE_ENDPOINT_ID_REMOVE = "page-remove";
 const PAGE_ENDPOINT_ID_SET = "page-set";
@@ -97,13 +100,13 @@ export function overrideCallback(
 			if (!isEmpty(query)) {
 				logger.info(`Saved workspaces requested with query: ${query}`);
 			}
-			logger.info(`Checking for custom workspace storage with endpoint id: ${WORKSPACE_ENDPOINT_ID_GET}`);
-			if (endpointProvider.hasEndpoint(WORKSPACE_ENDPOINT_ID_GET)) {
+			logger.info(`Checking for custom workspace storage with endpoint id: ${WORKSPACE_ENDPOINT_ID_LIST}`);
+			if (endpointProvider.hasEndpoint(WORKSPACE_ENDPOINT_ID_LIST)) {
 				logger.info("Requesting saved workspaces from custom storage");
 				const workspacesResponse = await endpointProvider.requestResponse<
-					{ query?: string },
+					{ platform: string; query?: string },
 					{ [key: string]: Workspace }
-				>(WORKSPACE_ENDPOINT_ID_GET, { query });
+				>(WORKSPACE_ENDPOINT_ID_LIST, { platform: fin.me.identity.uuid, query });
 
 				if (workspacesResponse) {
 					logger.info("Returning saved workspaces from custom storage");
@@ -129,10 +132,10 @@ export function overrideCallback(
 			logger.info(`Checking for custom workspace storage with endpoint id: ${WORKSPACE_ENDPOINT_ID_GET}`);
 			if (endpointProvider.hasEndpoint(WORKSPACE_ENDPOINT_ID_GET)) {
 				logger.info(`Requesting saved workspace from custom storage for workspace id: ${id}`);
-				const workspaceResponse = await endpointProvider.requestResponse<{ id: string }, Workspace>(
-					WORKSPACE_ENDPOINT_ID_GET,
-					{ id }
-				);
+				const workspaceResponse = await endpointProvider.requestResponse<
+					{ platform: string; id: string },
+					Workspace
+				>(WORKSPACE_ENDPOINT_ID_GET, { platform: fin.me.identity.uuid, id });
 				if (workspaceResponse) {
 					logger.info(`Returning saved workspace from custom storage for workspace id: ${id}`);
 					return workspaceResponse;
@@ -155,9 +158,13 @@ export function overrideCallback(
 			// in non-default location (e.g. on the server instead of locally)
 			logger.info(`Checking for custom workspace storage with endpoint id: ${WORKSPACE_ENDPOINT_ID_SET}`);
 			if (endpointProvider.hasEndpoint(WORKSPACE_ENDPOINT_ID_SET)) {
-				const success = await endpointProvider.action<{ id: string; payload: Workspace }>(
+				const success = await endpointProvider.action<{ platform: string; id: string; payload: Workspace }>(
 					WORKSPACE_ENDPOINT_ID_SET,
-					{ id: req.workspace.workspaceId, payload: req.workspace }
+					{
+						platform: fin.me.identity.uuid,
+						id: req.workspace.workspaceId,
+						payload: cleanupWorkspace(req.workspace)
+					}
 				);
 				if (success) {
 					logger.info(`Saved workspace with id: ${req.workspace.workspaceId} to custom storage`);
@@ -187,9 +194,13 @@ export function overrideCallback(
 			// in non-default location (e.g. on the server instead of locally)
 			logger.info(`Checking for custom workspace storage with endpoint id: ${WORKSPACE_ENDPOINT_ID_SET}`);
 			if (endpointProvider.hasEndpoint(WORKSPACE_ENDPOINT_ID_SET)) {
-				const success = await endpointProvider.action<{ id: string; payload: Workspace }>(
+				const success = await endpointProvider.action<{ platform: string; id: string; payload: Workspace }>(
 					WORKSPACE_ENDPOINT_ID_SET,
-					{ id: req.workspace.workspaceId, payload: req.workspace }
+					{
+						platform: fin.me.identity.uuid,
+						id: req.workspace.workspaceId,
+						payload: cleanupWorkspace(req.workspace)
+					}
 				);
 				if (success) {
 					logger.info(`Updated workspace with id: ${req.workspace.workspaceId} against custom storage`);
@@ -225,7 +236,10 @@ export function overrideCallback(
 			// in non-default location (e.g. on the server instead of locally)
 			logger.info(`Checking for custom workspace storage with endpoint id: ${WORKSPACE_ENDPOINT_ID_REMOVE}`);
 			if (endpointProvider.hasEndpoint(WORKSPACE_ENDPOINT_ID_REMOVE)) {
-				const success = await endpointProvider.action<{ id: string }>(WORKSPACE_ENDPOINT_ID_REMOVE, { id });
+				const success = await endpointProvider.action<{ platform: string; id: string }>(
+					WORKSPACE_ENDPOINT_ID_REMOVE,
+					{ platform: fin.me.identity.uuid, id }
+				);
 				if (success) {
 					logger.info(`Removed workspace with id: ${id} from custom storage`);
 				} else {
@@ -255,13 +269,13 @@ export function overrideCallback(
 			if (!isEmpty(query)) {
 				logger.info(`Saved pages requested with query: ${query}`);
 			}
-			logger.info(`Checking for custom page storage with endpoint id: ${PAGE_ENDPOINT_ID_GET}`);
-			if (endpointProvider.hasEndpoint(PAGE_ENDPOINT_ID_GET)) {
+			logger.info(`Checking for custom page storage with endpoint id: ${PAGE_ENDPOINT_ID_LIST}`);
+			if (endpointProvider.hasEndpoint(PAGE_ENDPOINT_ID_LIST)) {
 				logger.info("Getting saved pages from custom storage");
 				const pagesResponse = await endpointProvider.requestResponse<
-					{ query?: string },
+					{ platform: string; query?: string },
 					{ [key: string]: Page }
-				>(PAGE_ENDPOINT_ID_GET, { query });
+				>(PAGE_ENDPOINT_ID_LIST, { platform: fin.me.identity.uuid, query });
 				if (pagesResponse) {
 					logger.info("Returning saved pages from custom storage");
 					return Object.values(pagesResponse);
@@ -286,9 +300,10 @@ export function overrideCallback(
 			logger.info(`Checking for custom page storage with endpoint id: ${PAGE_ENDPOINT_ID_GET}`);
 			if (endpointProvider.hasEndpoint(PAGE_ENDPOINT_ID_GET)) {
 				logger.info(`Getting saved page from custom storage for page id: ${id}`);
-				const pageResponse = await endpointProvider.requestResponse<{ id: string }, Page>(
+				const pageResponse = await endpointProvider.requestResponse<{ platform: string; id: string }, Page>(
 					PAGE_ENDPOINT_ID_GET,
 					{
+						platform: fin.me.identity.uuid,
 						id
 					}
 				);
@@ -335,10 +350,14 @@ export function overrideCallback(
 			logger.info(`Checking for custom page storage with endpoint id: ${PAGE_ENDPOINT_ID_SET}`);
 			if (endpointProvider.hasEndpoint(PAGE_ENDPOINT_ID_SET)) {
 				logger.info(`Saving page with id: ${req.page.pageId} to custom storage`);
-				const success = await endpointProvider.action<{ id: string; payload: Page }>(PAGE_ENDPOINT_ID_SET, {
-					id: req.page.pageId,
-					payload: req.page
-				});
+				const success = await endpointProvider.action<{ platform: string; id: string; payload: Page }>(
+					PAGE_ENDPOINT_ID_SET,
+					{
+						platform: fin.me.identity.uuid,
+						id: req.page.pageId,
+						payload: cleanupPage(req.page)
+					}
+				);
 				if (success) {
 					logger.info(`Saved page with id: ${req.page.pageId} to custom storage`);
 				} else {
@@ -374,10 +393,14 @@ export function overrideCallback(
 			logger.info(`Checking for custom page storage with endpoint id: ${PAGE_ENDPOINT_ID_SET}`);
 			if (endpointProvider.hasEndpoint(PAGE_ENDPOINT_ID_SET)) {
 				logger.info(`Updating saved page and saving to custom storage with page id: ${req.page.pageId}`);
-				const success = await endpointProvider.action<{ id: string; payload: Page }>(PAGE_ENDPOINT_ID_SET, {
-					id: req.page.pageId,
-					payload: req.page
-				});
+				const success = await endpointProvider.action<{ platform: string; id: string; payload: Page }>(
+					PAGE_ENDPOINT_ID_SET,
+					{
+						platform: fin.me.identity.uuid,
+						id: req.page.pageId,
+						payload: cleanupPage(req.page)
+					}
+				);
 				if (success) {
 					logger.info(`Updated page with id: ${req.page.pageId} against custom storage`);
 				} else {
@@ -407,7 +430,10 @@ export function overrideCallback(
 			logger.info(`Checking for custom page storage with endpoint id: ${PAGE_ENDPOINT_ID_REMOVE}`);
 			if (endpointProvider.hasEndpoint(PAGE_ENDPOINT_ID_REMOVE)) {
 				logger.info(`deleting saved page from custom storage. PageId: ${id}`);
-				const success = await endpointProvider.action<{ id: string }>(PAGE_ENDPOINT_ID_REMOVE, { id });
+				const success = await endpointProvider.action<{ platform: string; id: string }>(
+					PAGE_ENDPOINT_ID_REMOVE,
+					{ platform: fin.me.identity.uuid, id }
+				);
 				if (success) {
 					logger.info(`Removed page with id: ${id} from custom storage`);
 				} else {