+++ how-to/workspace-platform-starter/CHANGELOG.md
@@ -7,6 +7,9 @@
 - Added support for OpenFin Workspace's Microsoft Low Code Integration (see [How To Setup Low Code Integrations](./docs/how-to-setup-low-code-integrations.md))
 - Fetch endpoints substitute url params for all http verbs
 - Fetch endpoints now support DELETE, PUT and PATCH http verbs
+- Page and Workspace storage using custom endpoints now reduces the JSON stored to just the essential data
+- Page and Workspace storage now separates the endpoints for getting all the entries `-list` and a single entry `-get`
+- Endpoint fetch requests now add the `Content-Type` and `Accept` headers with `application/json`
 
 ## v13